Authorization checks

+++ controllers/feed.js
@@ -2,6 +2,8 @@ const fs = require("fs");
 const { validationResult } = require("express-validator");
 
 const Post = require("../models/post");
+const User = require("../models/user");
+const user = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -15,13 +17,11 @@ exports.getPosts = (req, res, next) => {
         .skip((currentPage - 1) * perPage)
         .limit(perPage)
         .then((posts) => {
-          res
-            .status(200)
-            .json({
-              message: "Post Fetched Successfully",
-              posts: posts,
-              totalItems: totalItems,
-            });
+          res.status(200).json({
+            message: "Post Fetched Successfully",
+            posts: posts,
+            totalItems: totalItems,
+          });
         })
         .catch((err) => {
           if (!err.statusCode) {
@@ -55,20 +55,30 @@ exports.creatPosts = (req, res, next) => {
   // // for mac  const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   // send post to db
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "Destiny" },
+    creator: req.userId,
   });
   post
     .save()
     .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       console.log(result);
       res.status(201).json({
         message: "Post Created Successfully.",
-        post: result,
+        post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -124,6 +134,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized");
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl != post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -152,8 +167,20 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
       // Check login user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorized");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       Post.findByIdAndRemove(postId);
+    })
+    .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
     })
     .then((result) => {
       res.status(200).json({ message: "Post deleted successfully" });
